feat(as-3637): update the code to point to renamed folder full-appeal
