+++ packages/forms-web-app/__tests__/unit/routes/full-appeal/date-decision-due.test.js
@@ -1,18 +1,18 @@
 const { get, post } = require('../router-mock');
-const decisionDateController = require('../../../../src/controllers/full-planning/date-decision-due');
+const decisionDateController = require('../../../../src/controllers/full-appeal/date-decision-due');
 const fetchExistingAppealMiddleware = require('../../../../src/middleware/fetch-existing-appeal');
 const combineDateInputsMiddleware = require('../../../../src/middleware/combine-date-inputs');
 const { validationErrorHandler } = require('../../../../src/validators/validation-error-handler');
 const {
   rules: decisionDateValidationRules,
-} = require('../../../../src/validators/full-planning/date-decision-due');
+} = require('../../../../src/validators/full-appeal/date-decision-due');
 
-jest.mock('../../../../src/validators/full-planning/date-decision-due');
+jest.mock('../../../../src/validators/full-appeal/date-decision-due');
 
 describe('routes/eligibility/decision-date', () => {
   beforeEach(() => {
     // eslint-disable-next-line global-require
-    require('../../../../src/routes/full-planning/date-decision-due');
+    require('../../../../src/routes/full-appeal/date-decision-due');
   });
 
   it('should define the expected routes', () => {
