+++ packages/forms-web-app/src/routes/full-appeal/date-decision-due.js
@@ -1,12 +1,12 @@
 const express = require('express');
 
-const dateDecisionDueController = require('../../controllers/full-planning/date-decision-due');
+const dateDecisionDueController = require('../../controllers/full-appeal/date-decision-due');
 const fetchExistingAppealMiddleware = require('../../middleware/fetch-existing-appeal');
 const combineDateInputsMiddleware = require('../../middleware/combine-date-inputs');
 const { validationErrorHandler } = require('../../validators/validation-error-handler');
 const {
   rules: decisionDateDueValidationRules,
-} = require('../../validators/full-planning/date-decision-due');
+} = require('../../validators/full-appeal/date-decision-due');
 
 const router = express.Router();
